AI-Paper2Video/requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# AI/ML框架
torch==2.1.1
transformers==4.35.2
langchain==0.0.340
openai==1.3.0
anthropic==0.7.7

# 文档处理
pymupdf==1.23.8
python-multipart==0.0.6
pdf2image==1.16.3

# 图像处理
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# 视频处理
moviepy==1.0.3
ffmpeg-python==0.2.0

# 语音合成
elevenlabs==0.2.26
edge-tts==6.1.9

# 文件存储
boto3==1.34.0
minio==7.2.0

# 工具类
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
aiofiles==23.2.1

# 开发工具
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1